Make AbstractObservationProcess a TreeLikelihood
It has most of a TreeLikelihood's Inputs already, except under a slightly different field name (branchRateModeInput vs. branchRateModelInput being the most obnoxious one), so the other changes in this commit are making sure that the precise type of the Input (X vs. XInterface) match, and that no fields are duplicated (by name) any more.